+++ commands/moderation/purge/index.js
@@ -1,0 +1,6 @@
+import { SlashCommand } from '@aroleaf/djs-bot';
+
+export default new SlashCommand({
+  name: 'purge',
+  description: 'Commands for purging global messages',
+});