+++ commands/panic.js
@@ -15,6 +15,9 @@ export default new DJS.SlashCommand({
   if (global.bans.includes(interaction.user.id)) return reply('You are banned from this global channel.');
 
   await interaction.client.db.panic(global);
+  for (const channel of global.subscriptions.map(sub => interaction.client.channels.resolve(sub)).filter(c => c)) {
+    channel.global.panic();
+  }
   
   await interaction.client.channels.resolve(global.logs || process.env.LOGS)?.send({
     content: `<@&${process.env.PANIC_ROLE}>, ${interaction.user} enabled panic mode for ${global.name}. Once everything is stable again, use \`/unpanic\` to disable panic mode.`,